Times_Table
Write a program that outputs a times table from a given (user entered) integer value. It should start at 2 and output only the EVEN multiples. If the user entered the value 13, the even times table would be outputted in the following format:

The even timetable for 13 is: 
        2 times 13 is 26
        4 times 13 is 52
        6 times 13 is 78

This program should continue until the user chooses to exit. You should format this so that the text “times” and “is” is always in the same column. The program should output up to and including 20 times the user value.