Update and rename Stars_Diamond_Shape to Stars_Menu_All_Triangles_&_Diamond_Shape
In the sample code given for 'Iteration' in 1.1, there is a program for printing out a left-sided triangle of stars. 

Adapt the code to create a right-sided triangle.
Adapt the code to output a diamond.
Write a menu system so the user can choose the left or right-sided triangle, or diamond.
Enable the program to take user input for the symbol used to generate the pattern and dictate the size of the pattern.